+++ wow-webflux/src/main/kotlin/me/ahoo/wow/webflux/route/BatchResults.kt
@@ -13,19 +13,20 @@
 
 package me.ahoo.wow.webflux.route
 
+import io.github.oshai.kotlinlogging.KotlinLogging
 import me.ahoo.wow.api.modeling.AggregateId
 import me.ahoo.wow.exception.toErrorInfo
 import me.ahoo.wow.openapi.BatchResult
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 
-private val log = org.slf4j.LoggerFactory.getLogger("Wow.BatchResult")
+private val log = KotlinLogging.logger("Wow.BatchResult")
 
 fun Flux<AggregateId>.toBatchResult(afterId: String): Mono<BatchResult> {
     return this.materialize().reduce(BatchResult(afterId, 0)) { acc, signal ->
         if (signal.isOnError) {
-            if (log.isWarnEnabled) {
-                log.warn("Reduce onError.", signal.throwable)
+            log.warn(signal.throwable) {
+                "Reduce onError."
             }
             val error = signal.throwable!!.toErrorInfo()
             return@reduce acc.copy(
