+++ wow-webflux/src/test/kotlin/me/ahoo/wow/webflux/route/id/GlobalIdHandlerFunctionTest.kt
@@ -2,6 +2,7 @@ package me.ahoo.wow.webflux.route.id
 
 import me.ahoo.wow.openapi.context.OpenAPIComponentContext
 import me.ahoo.wow.openapi.global.GenerateGlobalIdRouteSpec
+import me.ahoo.wow.webflux.route.global.GlobalIdHandlerFunctionFactory
 import org.hamcrest.MatcherAssert
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.Test
